components/graph-service-backend-dev/pyproject.toml
[project]
name = "graph-service-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiobotocore>=3.0.0",
    "aiomqtt==2.3.0",
    "alembic==1.15.2",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.0",
    "asyncpg==0.31.0",
    "audit-lib>=0.1.0a2",
    "auth-lib>=0.1.0a3",
    "black==25.1.0",
    "boto3>=1.42.5",
    "certifi==2025.11.12",
    "charset-normalizer==3.4.4",
    "click==8.1.8",
    "ezdxf==1.4.3",
    "fastapi==0.124.0",
    "flake8==7.2.0",
    "geoalchemy2==0.17.1",
    "greenlet==3.3.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "idna==3.11",
    "igraph==0.11.8",
    "iniconfig==2.3.0",
    "loguru==0.7.3",
    "mako==1.3.10",
    "markupsafe==3.0.3",
    "matplotlib==3.10.8",
    "mccabe==0.7.0",
    "mypy-extensions==1.1.0",
    "numpy==2.1.3",
    "packaging==25.0",
    "paho-mqtt==2.1.0",
    "pathspec==0.12.1",
    "platformdirs==4.5.1",
    "pluggy==1.6.0",
    "psycopg2-binary==2.9.10",
    "pycodestyle==2.13.0",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.8.1",
    "pyflakes==3.3.2",
    "pygeohash==3.2.2",
    "pytest==8.3.5",
    "pytest-asyncio==0.25.3",
    "python-dotenv==1.1.0",
    "python-igraph==0.11.8",
    "python-multipart==0.0.20",
    "pyyaml==6.0.3",
    "redis==7.1.1",
    "requests==2.32.3",
    "sqlalchemy==2.0.44",
    "starlette==0.50.0",
    "texttable==1.7.0",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3==2.6.0",
    "uvicorn==0.34.2",
    "uvloop==0.22.1",
    "watchfiles==1.1.1",
    "websockets==15.0.1",
    "shapely>=2.1.2",
    "platform-sdk>=0.1.0a3",
]


[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest-cov>=7.0.0",
    "pytest-dotenv>=0.5.2",
    "ruff>=0.15.4",
    "testcontainers>=4.14.1",
]

[[tool.uv.index]]
name = "nexus"
url = "https://nexus.dmi-msk.ru/repository/pypi-dev-group/simple"

[tool.ruff]
line-length = 100
extend-exclude = [
    "src/migrations",
]
force-exclude = true

[tool.ruff.lint]
select = [
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # isort
    "I",
    # tidy imports
    "TID",
    # flake8-bugbear
    "B",
    # flake8-bandit
    "S",
    # pydocstyle
    "D",
    # pyugrade
    "UP"
]
ignore = ["D100", "D101", "D102", "D103", "D104", "D106", "D107", "D205", "D415", "D417", "B008", "E402", "E721", "W505"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "S105", "S106", "D100", "D101", "D102", "D103", "F403", "D106"]

[tool.ruff.lint.pycodestyle]
max-doc-length = 100
max-line-length = 100

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
docstring-code-format = true
